+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+
+## [0.4.1] - 2021-06-06
 ### Changed
 - Check allowed values for maximum charge current
 
@@ -49,7 +51,8 @@ Minor fix in observer interface
 ## [0.1.0] - 2021-05-26
 Initial release
 
-[unreleased]: https://github.com/tillsteinbach/WeConnect-python/compare/v0.4.0...HEAD
+[unreleased]: https://github.com/tillsteinbach/WeConnect-python/compare/v0.4.1...HEAD
+[0.4.1]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.4.1
 [0.4.0]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.4.0
 [0.3.2]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.3.2
 [0.3.1]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.3.1